initial

+++ lib/features/history/ui/pages/history_page.dart
@@ -6,12 +6,17 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:image_multi_type/image_multi_type.dart';
 import 'package:pump_app/core/extensions/extensions.dart';
+import 'package:pump_app/core/strings/app_color_manager.dart';
 import 'package:pump_app/core/widgets/app_bar/app_bar_widget.dart';
+import 'package:pump_app/core/widgets/my_button.dart';
+import 'package:pump_app/core/widgets/my_button.dart';
 import 'package:pump_app/core/widgets/my_card_widget.dart';
 
 import '../../../../core/util/my_style.dart';
+import '../../../../core/util/snack_bar_message.dart';
 import '../../../../router/app_router.dart';
 import '../../../db/models/app_specification.dart';
+import '../../../form/bloc/get_form_cubit/get_form_cubit.dart';
 import '../../bloc/get_history_cubit/get_history_cubit.dart';
 
 class HistoryPage extends StatelessWidget {
@@ -43,17 +48,38 @@ class HistoryPage extends StatelessWidget {
                 child: MyCardWidget(
                   child: Row(
                     children: [
+                      IconButton(
+                        onPressed: () {
+                          context
+                              .read<GetFormCubit>()
+                              .setQuestionsFromHistory(list: item.list);
+                          Navigator.pushNamed(context, RouteName.startForm);
+                        },
+                        icon: const ImageMultiType(
+                          url: Icons.edit,
+                          color: AppColorManager.ampere,
+                        ),
+                      ),
                       Expanded(
                         child: Column(
                           children: [
-                            DrawableText(text: item.date.formatDate),
+                            DrawableText(text: item.date.formatDateTime),
                             DrawableText(text: item.name ?? ''),
                           ],
                         ),
                       ),
                       IconButton(
                         onPressed: () {
-                          context.read<GetHistoryCubit>().deleteItem(i);
+                          showDialog(
+                            context: context,
+                            builder: (context) {
+                              return DeleteConfirmationDialog();
+                            },
+                          ).then((value) {
+                            if (value != null && value is bool && value) {
+                              context.read<GetHistoryCubit>().deleteItem(i);
+                            }
+                          });
                         },
                         icon: const ImageMultiType(
                           url: Icons.delete,
@@ -69,5 +95,59 @@ class HistoryPage extends StatelessWidget {
         },
       ),
     );
+  }
+}
+
+class DeleteConfirmationDialog extends StatelessWidget {
+  const DeleteConfirmationDialog({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: Text('تأكيد عملية الحذف'),
+      content: Text('هل أنت متأكد من حذف عنصر من السجل؟'),
+      actions: [
+        Row(
+          children: [
+            Expanded(
+              child: MyButton(
+                text: 'إلغاء',
+                color: Colors.black,
+                onTap: () {
+                  Navigator.of(context).pop(false); // Return false when cancel is pressed
+                },
+              ),
+            ),
+            10.0.horizontalSpace,
+            Expanded(
+              child: MyButton(
+                color: Colors.red,
+                text: 'حذف',
+                onTap: () {
+                  Navigator.of(context).pop(true); // Return true when delete is pressed
+                },
+              ),
+            ),
+          ],
+        ),
+      ],
+    );
+  }
+}
+
+// Usage example
+void showDeleteConfirmation(BuildContext context) async {
+  final bool deleteConfirmed = await showDialog(
+    context: context,
+    builder: (BuildContext context) {
+      return DeleteConfirmationDialog();
+    },
+  );
+
+  if (deleteConfirmed) {
+    // Delete logic goes here
+    print('Item deleted!');
+  } else {
+    print('Delete canceled!');
   }
 }
